refactor(facade): return function instead of running it
Return the function that executes a given command instead of simply
executing it if it is found. This is done for the sake of clarity and
single responsibility: should a function called 'parse' yield the result
of said parsing or DO something as a result of said parsing?

IMHO, this improves legibility.