feat: list available commands in usage message
List the available commands in the usage message by turning the
CommandFacade instance into a global variable, initialized before
main(), and accessing the list_commands() function from it.

To ensure a possible destructor method is called upon exit, an atexit
handler (delete_facade) was introduced that invokes the delete
operation.